clean up the code; some fixes
