+++ src/main/java/com/faforever/client/chat/test/ListItem.java
@@ -6,9 +6,16 @@ import com.faforever.client.theme.UiService;
 import javafx.scene.Node;
 import org.fxmisc.flowless.Cell;
 
+import java.util.Comparator;
+import java.util.Locale;
 import java.util.Optional;
 
 public abstract class ListItem {
+
+  public static Comparator<ListItem> getComparator() {
+    return Comparator.comparing(ListItem::getCategory).thenComparing(item -> item.getUser()
+        .map(user -> user.getUsername().toLowerCase(Locale.ROOT)).orElse(""));
+  }
 
   public abstract Cell<ListItem, Node> createCell(UiService uiService);
